% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{add_chars}
\alias{add_chars}
\title{add_chars}
\usage{
add_chars(obj, howmuch = 0, where = "fist", what = "0")
}
\arguments{
\item{obj}{formatG object}

\item{howmuch}{Number of characters}

\item{where}{"start" to add characters at the beginning of the string,
"end" to add characters at the end of the string}

\item{what}{Character to add}
}
\description{
Add characters at the beggining/end of each cell of the matrix
referenced in a formatG object
}
