% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{alleles_to_dummy}
\alias{alleles_to_dummy}
\title{alleles_to_dummy}
\usage{
alleles_to_dummy(obj, NA_action = "NA")
}
\arguments{
\item{obj}{formatG object}

\item{NA_action}{The NA action ("0" or "NA") for recoding NAs.}
}
\description{
Converts a matrix of alleles into a matrix of dummy factors
}
