% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{alleles_to_loci}
\alias{alleles_to_loci}
\title{alleles_to_loci}
\usage{
alleles_to_loci(obj, sep = "")
}
\arguments{
\item{obj}{formatG object}
}
\description{
Creates a individuals x 1 matrix,  with a loci in a column,
from a two-column matrix with the alleles
}
