% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\docType{methods}
\name{$,formatG-method}
\alias{$,formatG-method}
\title{$}
\usage{
\S4method{$}{formatG}(x, name)
}
\arguments{
\item{x}{formatG object}

\item{name}{Name of the method}
}
\description{
Method accessor
}
\keyword{internal}
