% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{formatter_chunk}
\alias{formatter_chunk}
\title{formatter_chunk}
\usage{
formatter_chunk(dbR6_object, path, formatG_input = NULL, ploidy, format_fun,
  out_df, has_rownames = TRUE, has_colnames = TRUE, chunksize = 1000L,
  token = ".", NA_action = "NA", sep_loc = " ", howmuch = 0,
  where = "last", what = "0", return_formatG = FALSE,
  input_or_output = c("data_input", "dataS", "dataI"), set_ncol = 0, ...)
}
\description{
formatter_chunk
}
