% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{formatter_chunk_from_db}
\alias{formatter_chunk_from_db}
\title{formatter_chunk_from_db}
\usage{
formatter_chunk_from_db(dbR6_object, table, formatG_input = NULL, ploidy,
  format_fun, out_df, has_rownames = TRUE, chunksize = 1000L, token = ".",
  NA_action = "NA", howmuch = 0, where = "last", what = "0",
  default_levels = character(0), return_formatG = FALSE,
  input_or_output = c("data_input", "dataS", "dataI"), set_ncol = 0, ...)
}
\description{
formatter_chunk_from_db
}
