% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_db.R
\name{genetic_db}
\alias{genetic_db}
\title{Generate a genetic database}
\usage{
genetic_db(G = data.frame(), type = c("codominant", "dominant"),
  ploidy = 2, sep = "", n_loci = 2, ncod = 1, missing = c("0", "NA"),
  NA.char = "NA", poly.level = 5, default_levels = character(0),
  chunksize = 1000, has_rownames = TRUE, has_colnames = TRUE,
  char_per_allele = 1, store_G = FALSE)
}
\description{
Generate a genetic database
}
