% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\docType{methods}
\name{initialize,formatG-method}
\alias{initialize,formatG-method}
\title{formatG initializer}
\usage{
\S4method{initialize}{formatG}(.Object, data, ploidy, default_levels, obj_ncol,
  char_per_allele, has_rownames, has_colnames, colnames_vector)
}
\description{
formatG initializer
}
\keyword{internal}
