% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{loci_to_alleles}
\alias{loci_to_alleles}
\title{loci_to_alleles}
\usage{
loci_to_alleles(obj)
}
\arguments{
\item{input}{individuals x 1 StringMatrix, with locus in column}
}
\description{
Split locus in column into alleles in columns
}
