% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{loci_token_to_alleles}
\alias{loci_token_to_alleles}
\title{loci_token_to_alleles}
\usage{
loci_token_to_alleles(obj, token = ".")
}
\arguments{
\item{obj}{formatG object}
}
\description{
Split locus in column with alleles separated by a character into alleles in columns
}
