% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatG.R
\name{process_chunk}
\alias{process_chunk}
\title{process chunk}
\usage{
process_chunk(obj, input_data, fun_name, has_rownames = TRUE, token = ".",
  NA_action = "NA", howmuch = 0, where = "last", what = "0",
  input_or_output = c("data_input", "dataS", "dataI"), set_ncol = 0)
}
\arguments{
\item{input}{individuals x 1 StringMatrix, with locus in column}
}
\description{
Split locus in column into alleles in columns
}
